Add function data to SummaryTable
Now I need to fix a few things. For one, the default receptor names need to be used as the dictionary entries in the summary field, meaning for a given sheet, the receptor name has to be properly matched. This information may be in the sheet name, in the file name, or both, and is usually written at the top of the spreadsheet as well. Furthermore, the drug names need to be identified properly.