Add function summary table method
The basic structure of the binding and function tables have been established. Now I need to fix the multiple issues that have arisen due to differential formatting of the data from different receptors. The names of drugs and receptors are not always correct and the values are sometimes incorrect or missing