Add draggable cell selection
It's currently bound to the left-mouse drag at all times in the CustomTable class, even though the callers need to be able to specify when it is bound. I'll need to figure out a better solution, this just implements the cell selection at all times which conflicts with the row or triplet selection in the various TriplicateGUI tables.