Begin strategy object refactor
It is functioning as before, with rows and triplets selectable. It's almost certainly broken for the original purpose but I will return to that later. I need to remove all mentions of row/triplet selection from the table itself as those are now injected using the style and selection strategy objects