Initialize SQL database schema
No SQL has been implemented yet, but this is the start of the porting from h5 to SQL storage of plate data. The PlateData class in bindMods will be used to store metadata and count data for each well and can be applied to any assay type. The wells will individually be stored as rows in the SQL database.